fogbugz:13393
updated components in the Intel folder to allow for the successful tearing
down of an Intel-based DirectShow graph before any samples were pushed
through it. Previously, upon teardown of such a graph some components were
uninitialized, which went ignored so as to dereference null pointers
throughout. This is what was directly causing the crash reported in
fogbugz:13393